Re-use expectations inside bnToU8a
